type migration of latitude and longitude

+++ src/main/java/com/mstiehr/model/Position.java
@@ -12,9 +12,9 @@ public class Position
     private long id;
 
     @Column(name = "longitude")
-    private long longitude;
+    private double longitude;
     @Column(name = "latitude")
-    private long latitude;
+    private double latitude;
 
     public Position() {
     }
@@ -24,19 +24,19 @@ public class Position
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
